requirements.txt
mkdocs==1.0.3
gitpython
requests